% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{validate_imputation}
\alias{validate_imputation}
\title{Validation Protocol}
\usage{
validate_imputation(
  DT,
  vars,
  id_var,
  time_var,
  impute_fn,
  mask_rate = 0.2,
  seed = 123,
  plot_worst_n = 4,
  ...
)
}
\arguments{
\item{DT}{Data.table.}

\item{vars}{Vector of variable names.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{impute_fn}{The function to test (e.g., kalman_fast).}

\item{mask_rate}{Fraction of data to mask (0-1).}

\item{seed}{Random seed.}

\item{plot_worst_n}{Integer. Number of diagnostic plots to return.}

\item{...}{Arguments passed to impute_fn.}
}
\description{
Masks a percentage of data and tests the accuracy of imputation functions.
}
