% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastImpute-package.R
\docType{package}
\name{fastImpute-package}
\alias{fastImpute}
\alias{fastImpute-package}
\title{fastImpute: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
