% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{fbi_fast}
\alias{fbi_fast}
\title{Fast Factor-Based Imputation}
\usage{
fbi_fast(
  DT,
  vars,
  id_var,
  time_var,
  rank = 4L,
  thresh = 1e-05,
  max_iter = 20L,
  out_suff = "_fbi",
  max_hole = Inf,
  max_endpoint = Inf,
  verbose = TRUE
)
}
\arguments{
\item{DT}{A data.table.}

\item{vars}{Character vector.}

\item{id_var}{Character string.}

\item{time_var}{Character string.}

\item{rank}{Integer.}

\item{thresh}{Numeric.}

\item{max_iter}{Integer.}

\item{out_suff}{Character string or NULL.}

\item{max_hole}{Integer or Inf.}

\item{max_endpoint}{Integer or Inf.}

\item{verbose}{Logical.}
}
\description{
Imputes missing values using an EM-SVD (Factor) approach, similar to Bai & Ng (2019),
optimized for large panels.
}
